groupe-48-IA explicable pour décisions d'investissement (XAI Finance)/requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
seaborn
requests
python-dotenv
yfinance
ta
pyarrow
pyarrow
fastapi
uvicorn
python-multipart
yahooquery
